Complete functions _.uniq and _.map
All tests pass for _.uniq and _.map. I added a declaring sorting capability for _.uniq, although I don't think it works the way the normal _.uniq works. Additionally, although the test passes, there is no way for an iterator to interact with _.uniq. I'm still experimenting with whether the iterator should execute with each unique value pushed or with every value in the array.