Fix _.uniq to accomodate an iterator
Previously didn't understand what the iterator did. Did some research and found out that the iterator in a _.uniq statement compares the results of putting in an arrays values through the iterator to make sure they are unique, but you still push the original values into the resulting array. These changes reflect that understanding.